+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 1.6.0 (?)
 Changes in this release:
+- Added the ability to assert the expected 'change' of a deploy
 
 # v 1.5.0 (2021-03-26)
 Changes in this release: